use shorts instead of ints in blocksets

+++ maplesat/core/Solver.h
@@ -49,8 +49,8 @@ public:
 
     char unit_clauses[200000];
     bool addunits = false;
-    std::set<std::array<int, 36>> blockset[5];
-    std::set<std::array<int, 36>> blockset2[5];
+    std::set<std::array<short, 36>> blockset[5];
+    std::set<std::array<short, 36>> blockset2[5];
     double nautytime = 0;
     double lookuptime = 0;
